Make __undefined__ processing recursive for nested objects
