+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "puty",
-  "version": "0.0.5-rc1",
+  "version": "0.0.6",
   "description": "A tooling function to test javascript functions and classes.",
   "main": "src/index.js",
   "type": "module",
